\name{print.RV}
\alias{print.RV}
\title{Print random variable}
\usage{
  \method{print}{RV} (x, ..., digits = 4)
}
\arguments{
  \item{x}{A random variable}

  \item{digits}{The number of digits to use when printing}

  \item{\ldots}{Further parameters to pass to
  print.data.frame}
}
\description{
  Provide a nice way to view a random variable
}

